refactor: 게시물 엔티티(Post.java)의 디렉토리 이동
